--- a/NodeReact.Benchmarks/Program.cs
+++ b/NodeReact.Benchmarks/Program.cs
@@ -17,8 +17,8 @@ namespace NodeReact.Benchmarks
             //}
 
 
-            //BenchmarkRunner.Run<SingleComponentBenchmark>();
-            BenchmarkRunner.Run<WebSimulateBenchmark>();
+            BenchmarkRunner.Run<SingleComponentBenchmark>();
+            //BenchmarkRunner.Run<WebSimulateBenchmark>();
             Console.ReadKey();
 		}
 	}