--- a/NodeReact.Benchmarks/Program.cs
+++ b/NodeReact.Benchmarks/Program.cs
@@ -18,7 +18,7 @@ namespace NodeReact.Benchmarks
 
 
             BenchmarkRunner.Run<SingleComponentBenchmark>();
-            //BenchmarkRunner.Run<WebSimulateBenchmark>();
+           // BenchmarkRunner.Run<WebSimulateBenchmark>();
             Console.ReadKey();
 		}
 	}