--- a/NodeReact.Benchmarks/SingleComponentBenchmark.cs
+++ b/NodeReact.Benchmarks/SingleComponentBenchmark.cs
@@ -3,14 +3,12 @@ using BenchmarkDotNet.Attributes;
 using Microsoft.Extensions.DependencyInjection;
 using NodeReact.Components;
 using React;
-using ZeroReactComponent = ZeroReact.Components.ReactComponent;
-using NodeReactComponent = NodeReact.Components.ReactComponent;
 
 namespace NodeReact.Benchmarks
 {
-	public class SingleComponentBenchmark : BaseBenchmark
-	{
-		private readonly NoTextWriter tk = new NoTextWriter();
+    public class SingleComponentBenchmark : BaseBenchmark
+    {
+        private readonly NoTextWriter tk = new NoTextWriter();
 
         [Benchmark]
         public async Task NodeReact_RenderRouterSingle()
@@ -19,7 +17,7 @@ namespace NodeReact.Benchmarks
             {
                 var reactContext = scope.ServiceProvider.GetRequiredService<NodeReact.IReactScopedContext>();
 
-                var component = reactContext.CreateComponent<ReactRouterComponent>("__desktopComponents.App");
+                var component = reactContext.CreateComponent<NodeReact.Components.ReactRouterComponent>("__desktopComponents.App");
                 component.Props = _testData;
                 component.ServerOnly = true;
                 component.Path = "/movie/246436/";
@@ -37,7 +35,7 @@ namespace NodeReact.Benchmarks
             {
                 var reactContext = scope.ServiceProvider.GetRequiredService<NodeReact.IReactScopedContext>();
 
-                var component = reactContext.CreateComponent<NodeReactComponent>("__components.MovieAboutPage");
+                var component = reactContext.CreateComponent<NodeReact.Components.ReactComponent>("__components.MovieAboutPage");
                 component.Props = _testData;
                 component.ServerOnly = true;
 
@@ -47,14 +45,14 @@ namespace NodeReact.Benchmarks
             }
         }
 
-       // [Benchmark]
+        [Benchmark]
         public async Task ZeroReact_RenderSingle()
         {
             using (var scope = sp.CreateScope())
             {
                 var reactContext = scope.ServiceProvider.GetRequiredService<ZeroReact.IReactScopedContext>();
 
-                var component = reactContext.CreateComponent<ZeroReactComponent>("__components.MovieAboutPage");
+                var component = reactContext.CreateComponent<ZeroReact.Components.ReactComponent>("__components.MovieAboutPage");
                 component.Props = _testData;
                 component.ServerOnly = true;
 
@@ -64,14 +62,32 @@ namespace NodeReact.Benchmarks
             }
         }
 
-      //  [Benchmark]
-	    public void ReactJs_RenderSingle()
-	    {
-		    var environment = AssemblyRegistration.Container.Resolve<IReactEnvironment>();
-		    var component = environment.CreateComponent("__components.MovieAboutPage", _testData, serverOnly: true);
+        [Benchmark]
+        public async Task ZeroReact_RenderRouterSingle()
+        {
+            using (var scope = sp.CreateScope())
+            {
+                var reactContext = scope.ServiceProvider.GetRequiredService<ZeroReact.IReactScopedContext>();
 
-		    component.RenderHtml(tk, renderServerOnly: true);
-		    environment.ReturnEngineToPool();
+                var component = reactContext.CreateComponent<ZeroReact.Components.ReactRouterComponent>("__desktopComponents.App");
+                component.Props = _testData;
+                component.ServerOnly = true;
+                component.Path = "/movie/246436/";
+
+                await component.RenderRouterWithContext();
+
+                component.WriteOutputHtmlTo(tk);
+            }
+        }
+
+        [Benchmark]
+        public void ReactJs_RenderSingle()
+        {
+            var environment = AssemblyRegistration.Container.Resolve<IReactEnvironment>();
+            var component = environment.CreateComponent("__components.MovieAboutPage", _testData, serverOnly: true);
+
+            component.RenderHtml(tk, renderServerOnly: true);
+            environment.ReturnEngineToPool();
         }
     }
 }