--- a/NodeReact/Jering.Javascript.NodeJS/InvocationContent.cs
+++ b/NodeReact/Jering.Javascript.NodeJS/InvocationContent.cs
@@ -14,7 +14,7 @@ namespace NodeReact
     {
         private static readonly JsonSerializerOptions jsonSerializerOptions = new JsonSerializerOptions
         {
-            Converters = { new CustomSerializerDeserializer() },
+            Converters = { new MemoryOwnerJsonConverter() },
             Encoder = JavaScriptEncoder.UnsafeRelaxedJsonEscaping,
             DefaultBufferSize = 64536,