--- a/NodeReact/Jering.Javascript.NodeJS/NodeInvocationError.cs
+++ b/NodeReact/Jering.Javascript.NodeJS/NodeInvocationError.cs
@@ -1,0 +1,11 @@
+﻿using System.Text.Json.Serialization;
+
+namespace NodeReact
+{
+    public class NodeInvocationError
+    {
+        public string ErrorMessage { get; set; }
+
+        public string ErrorStack { get; set; }
+    }
+}