… NodeReact/ServiceCollectionExtensions.cs
… NodeReact/Utils/PooledCharBuffer.cs