**Before**
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.17763.475 (1809/October2018Update/Redstone5)
Intel Core i7-7700 CPU 3.60GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100
  [Host]     : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), X64 RyuJIT
  DefaultJob : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), X64 RyuJIT


```
|                       Method |     Mean |    Error |   StdDev |   Median | Gen 0 | Gen 1 | Gen 2 |  Allocated |
|----------------------------- |---------:|---------:|---------:|---------:|------:|------:|------:|-----------:|
| NodeReact_RenderRouterSingle | 26.34 ms | 0.753 ms | 2.173 ms | 26.17 ms |     - |     - |     - | 2097.53 KB |
|       NodeReact_RenderSingle | 25.63 ms | 0.768 ms | 2.216 ms | 25.37 ms |     - |     - |     - | 1262.37 KB |
|       ZeroReact_RenderSingle | 50.80 ms | 2.594 ms | 7.316 ms | 50.28 ms |     - |     - |     - |    4.73 KB |
| ZeroReact_RenderRouterSingle | 52.67 ms | 1.792 ms | 4.994 ms | 51.27 ms |     - |     - |     - |     7.4 KB |
|         ReactJs_RenderSingle | 54.38 ms | 1.899 ms | 5.599 ms | 54.82 ms |     - |     - |     - | 1305.15 KB |


**After**
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.17763.475 (1809/October2018Update/Redstone5)
Intel Core i7-7700 CPU 3.60GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100
  [Host]     : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), X64 RyuJIT
  DefaultJob : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), X64 RyuJIT


```
|                       Method |     Mean |    Error |   StdDev |   Median | Gen 0 | Gen 1 | Gen 2 |  Allocated |
|----------------------------- |---------:|---------:|---------:|---------:|------:|------:|------:|-----------:|
| NodeReact_RenderRouterSingle | 25.79 ms | 0.847 ms | 2.472 ms | 25.62 ms |     - |     - |     - |   10.09 KB |
|       NodeReact_RenderSingle | 25.91 ms | 0.745 ms | 2.149 ms | 25.63 ms |     - |     - |     - |   10.38 KB |
|       ZeroReact_RenderSingle | 50.92 ms | 2.717 ms | 7.707 ms | 49.87 ms |     - |     - |     - |    4.73 KB |
| ZeroReact_RenderRouterSingle | 53.73 ms | 2.254 ms | 6.393 ms | 51.96 ms |     - |     - |     - |     7.4 KB |
|         ReactJs_RenderSingle | 55.61 ms | 1.890 ms | 5.571 ms | 56.25 ms |     - |     - |     - | 1305.15 KB |

@JeremyTCD